+++ src/gromo/modules/linear_growing_module.py
@@ -1043,25 +1043,35 @@ class LinearGrowingModule(GrowingModule):
 
         return alpha_weight, alpha_bias, omega, self.eigenvalues_extension
 
-    @staticmethod
-    def get_fan_in_from_layer(layer: torch.nn.Linear) -> int:  # type: ignore
+    def get_fan_in_from_layer(  # type: ignore
+        self, layer: torch.nn.Linear | None = None, num_neurons: int | None = None
+    ) -> int:
         """
-        Get the fan_in (number of input features) from a given layer.
+        Get the fan_in (number of input features) from a given layer
+            or from a given number of neurons.
 
         Parameters
         ----------
-        layer: torch.nn.Linear
+        layer: torch.nn.Linear | None
             layer to get the fan_in from
+        num_neurons: int | None
+            number of neurons in the layer
 
         Returns
         -------
         int
             fan_in of the layer
         """
-        assert isinstance(layer, torch.nn.Linear), (
-            f"The layer should be a torch.nn.Linear but got {type(layer)}."
-        )
-        return layer.in_features
+        if layer is not None:
+            assert isinstance(layer, torch.nn.Linear), (
+                f"The layer should be a torch.nn.Linear but got {type(layer)}."
+            )
+            return layer.in_features
+        else:
+            assert num_neurons is not None, (
+                "Either layer or num_neurons should be provided."
+            )
+            return num_neurons
 
     def create_layer_in_extension(self, extension_size: int) -> None:
         """